refactor: separate course search and planner features
