+++ features/course-search/course-modal-provider.tsx
