+++ features/course-search/course-search-panel.tsx
